# openapi_paths.yaml
openapi: 3.0.0
info:
  title: Bar Management API
  version: '0.1'
paths:
  /ingredients:
    get:
      summary: List ingredients
      parameters:
        - in: query
          name: skip
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
    post:
      summary: Create ingredient
  /recipes/search:
    get:
      summary: Search recipes from CocktailDB
      parameters:
        - in: query
          name: q
          schema:
            type: string
  /search:
    get:
      summary: Search local recipes
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: available_only
          schema:
            type: boolean
        - in: query
          name: order_missing
          schema:
            type: boolean
        - in: query
          name: skip
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
  /recipes:
    get:
      summary: List recipes
      parameters:
        - in: query
          name: skip
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
    post:
      summary: Add recipe from CocktailDB
  /recipes/{recipe_id}:
    get:
      summary: Get recipe with inventory
      parameters:
        - in: path
          name: recipe_id
          required: true
          schema:
            type: integer
...